+++ src/ShaderManager.cpp
@@ -5,18 +5,18 @@ using namespace warp;
 void ShaderManager::setActive(Shader::ID id)
 {
     if (activeID == id) return;
-    shaders.at(static_cast<long>(id))->bind();
+    shaders.at(static_cast<size_t>(id))->bind();
     activeID = id;
 }
 
 std::shared_ptr<Shader> ShaderManager::getActive()
 {
-    return shaders.at(static_cast<long>(activeID));
+    return shaders.at(static_cast<size_t>(activeID));
 }
 
 std::shared_ptr<Shader> ShaderManager::get(Shader::ID id)
 {
-    return shaders.at(static_cast<long>(id));
+    return shaders.at(static_cast<size_t>(id));
 }
 
 Shader::ID ShaderManager::add(std::shared_ptr<Shader> shader)
